feat: rename and organize unix wrappers
