--- src/data_provider/src/extended_sources/groups/include/group_wrapper_linux.hpp
@@ -1,13 +1,0 @@
-#include "igroup_wrapper_linux.hpp"
-
-#include <grp.h>
-
-class GroupWrapperLinux : public IGroupWrapperLinux
-{
-    public:
-
-        int getgrouplist(const char* user, gid_t group, gid_t* groups, int* ngroups) const override
-        {
-            return ::getgrouplist(user, group, groups, ngroups);
-        }
-};
