--- src/data_provider/src/extended_sources/groups/include/igroup_wrapper_linux.hpp
@@ -1,11 +1,0 @@
-#pragma once
-
-#include <grp.h>
-
-class IGroupWrapperLinux
-{
-    public:
-        virtual ~IGroupWrapperLinux() = default;
-
-        virtual int getgrouplist(const char* user, gid_t group, gid_t* groups, int* ngroups) const = 0;
-};
