--- src/data_provider/src/extended_sources/groups/include/iopen_directory_utils_wrapper.hpp
@@ -1,13 +1,0 @@
-#pragma once
-
-#include <map>
-#include <string>
-
-class IODUtilsWrapper
-{
-    public:
-        virtual ~IODUtilsWrapper() = default;
-        virtual void genEntries(const std::string& recordType,
-                                const std::string* record,
-                                std::map<std::string, bool>& names) = 0;
-};
