--- src/data_provider/src/extended_sources/groups/include/isystem_wrapper.hpp
@@ -1,9 +1,0 @@
-#pragma once
-
-class ISystemWrapper
-{
-    public:
-        virtual ~ISystemWrapper() = default;
-
-        virtual long sysconf(int name) const = 0;
-};
