--- src/data_provider/src/extended_sources/groups/include/system_wrapper.hpp
@@ -1,12 +1,0 @@
-#include "isystem_wrapper.hpp"
-
-#include <unistd.h>
-
-class SystemWrapper : public ISystemWrapper
-{
-    public:
-        long sysconf(int name) const override
-        {
-            return ::sysconf(name);
-        }
-};
