+++ src/data_provider/src/extended_sources/groups/include/user_groups_darwin.hpp
@@ -1,10 +1,19 @@
+/* Copyright (C) 2015, Wazuh Inc.
+ * All rights reserved.
+ *
+ * This program is free software; you can redistribute it
+ * and/or modify it under the terms of the GNU General Public
+ * License (version 2) as published by the FSF - Free Software
+ * Foundation.
+ */
+
 #pragma once
 
 #include <set>
 
 #include "json.hpp"
-#include "igroup_wrapper_darwin.hpp"
-#include "ipasswd_wrapper_darwin.hpp"
+#include "igroup_wrapper.hpp"
+#include "ipasswd_wrapper.hpp"
 #include "iopen_directory_utils_wrapper.hpp"
 
 #define EXPECTED_GROUPS_MAX 64
