+++ src/data_provider/src/extended_sources/groups/src/user_groups_darwin.cpp
@@ -1,7 +1,16 @@
+/* Copyright (C) 2015, Wazuh Inc.
+ * All rights reserved.
+ *
+ * This program is free software; you can redistribute it
+ * and/or modify it under the terms of the GNU General Public
+ * License (version 2) as published by the FSF - Free Software
+ * Foundation.
+ */
+
 #include <iostream>
 #include "user_groups_darwin.hpp"
-#include "group_wrapper_darwin.hpp"
-#include "passwd_wrapper_darwin.hpp"
+#include "group_wrapper.hpp"
+#include "passwd_wrapper.hpp"
 #include "open_directory_utils_wrapper.hpp"
 
 UserGroupsProvider::UserGroupsProvider(std::shared_ptr<IGroupWrapperDarwin> groupWrapper,
