+++ src/data_provider/src/extended_sources/wrappers/unix/darwin/group_wrapper.hpp
@@ -1,4 +1,15 @@
-#include "igroup_wrapper_darwin.hpp"
+/* Copyright (C) 2015, Wazuh Inc.
+ * All rights reserved.
+ *
+ * This program is free software; you can redistribute it
+ * and/or modify it under the terms of the GNU General Public
+ * License (version 2) as published by the FSF - Free Software
+ * Foundation.
+ */
+
+#pragma once
+
+#include "igroup_wrapper.hpp"
 
 #include <grp.h>
 #include <unistd.h>
