+++ src/data_provider/src/extended_sources/wrappers/unix/darwin/igroup_wrapper.hpp
@@ -1,3 +1,12 @@
+/* Copyright (C) 2015, Wazuh Inc.
+ * All rights reserved.
+ *
+ * This program is free software; you can redistribute it
+ * and/or modify it under the terms of the GNU General Public
+ * License (version 2) as published by the FSF - Free Software
+ * Foundation.
+ */
+
 #pragma once
 
 #include <grp.h>
